requirements.txt
requests
itsdangerous==2.2.0

xlrd==2.0.1
reportlab==4.1.0
gunicorn==23.0.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9

Pillow
requests
openpyxl
xlsxwriter
pandas==2.2.2
